feat: custom exception implementation

+++ LeaveManagmentClean.Application/Features/LeaveType/Commands/DeleteLeaveType/DeleteLeaveTypeHandler.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using LeaveManagmentClean.Application.Contracts.Persistence;
+using LeaveManagmentClean.Application.Exceptions;
 using MediatR;
 
 namespace LeaveManagmentClean.Application.Features.LeaveType.Commands.DeleteLeaveType
@@ -17,6 +18,11 @@ namespace LeaveManagmentClean.Application.Features.LeaveType.Commands.DeleteLeav
         {
             var leaveTypeToDelete = await _leaveTypeRepository.GetByIdAsync(request.Id);
 
+            if (leaveTypeToDelete is null)
+            {
+                throw new NotFoundException(nameof(LeaveType), request.Id);
+            }
+            
             await _leaveTypeRepository.DeleteAsync(leaveTypeToDelete);
 
             return Unit.Value;
